fix: render email body html + attachments in job view (#212)
The job view's sent-email row showed body_text only and never joined the
attachments table, so paragraph spacing was flattened and sent PDFs were
invisible. Mirror the inbox reader by sharing the EmailBodyFrame and
EmailAttachments components and joining email_attachments on the fetch.

Also route the compose body_text derivation through htmlToText() instead
of DOM textContent so the stored plain-text body and the SMTP `text`
field retain paragraph spacing for any text-only client.

Live UI not verified in this remote env (no SMTP creds); confirmed via
static analysis against the working inbox renderer and a regression test
pinning the htmlToText paragraph + line-break semantics.
